Revert "Merge branch 'main' into feat/AI-Assistant"
This reverts commit 803cb4e5822cca022d062030b7ce4d9bb4032adf, reversing
changes made to 6d4f844ff953c77deb43aedab43eeaecfdb154af.
